+++ packages/storybook/config/preview-head.html
@@ -1,1 +1,2 @@
-<link rel="icon" type="image/svg+xml" href="favicon.svg" />
+<link rel="preconnect" href="https://cdn.fonts.net">
+<link href="https://cdn.fonts.net/kit/98d15e86-0ba4-4c49-97a9-a796340ef3f3/98d15e86-0ba4-4c49-97a9-a796340ef3f3.css" rel="stylesheet" />
