+++ packages/storybook/package.json
@@ -20,12 +20,6 @@
     "storybook": "storybook dev --config-dir config/ --port 6006"
   },
   "devDependencies": {
-    "@tilburg/assets": "workspace:*",
-    "@tilburg/components-css": "workspace:*",
-    "@tilburg/design-tokens": "workspace:*",
-    "@tilburg/font": "workspace:*",
-    "@tilburg/web-components-react": "workspace:*",
-    "@tilburg/web-components-stencil": "workspace:*",
     "@storybook/addon-a11y": "8.0.9",
     "@storybook/addon-actions": "8.0.9",
     "@storybook/addon-docs": "8.0.9",
@@ -34,8 +28,16 @@
     "@storybook/addon-viewport": "8.0.9",
     "@storybook/react": "8.0.9",
     "@storybook/react-vite": "8.0.9",
+    "@tilburg/assets": "workspace:*",
+    "@tilburg/components-css": "workspace:*",
+    "@tilburg/design-tokens": "workspace:*",
+    "@tilburg/font": "workspace:*",
+    "@tilburg/web-components-react": "workspace:*",
+    "@tilburg/web-components-stencil": "workspace:*",
     "@types/react": "18.3.1",
     "@types/react-dom": "18.3.0",
+    "@utrecht/component-library-react": "3.0.1-alpha.33",
+    "@utrecht/components": "3.0.1-alpha.33",
     "react": "18.3.1",
     "react-dom": "18.3.1",
     "rimraf": "5.0.5",
