+++ packages/storybook/src/button.md
@@ -1,0 +1,5 @@
+<!-- @license CC0-1.0 -->
+
+# Button
+
+Button component for Tilburg
