--- packages/storybook/src/css-button.stories.tsx
@@ -1,48 +1,0 @@
-/* @license CC0-1.0 */
-
-import type { Meta, StoryObj } from '@storybook/react';
-import readme from '@tilburg/components-css/button/README.md?raw';
-import { PropsWithChildren } from 'react';
-import '@tilburg/components-css/button/index.scss';
-
-const Button = ({ children }: PropsWithChildren<{}>) => (
-  <button className="tilburg-button" type="button">
-    <span className="tilburg-button__text">{children}</span>
-  </button>
-);
-
-const meta = {
-  title: 'CSS Component/Button',
-  id: 'css-button',
-  component: Button,
-  argTypes: {
-    children: {
-      name: 'Content',
-      description: 'Button text',
-      type: {
-        name: 'string',
-        required: true,
-      },
-      defaultValue: '',
-    },
-  },
-  args: {
-    children: 'Opslaan en verder',
-  },
-  tags: ['autodocs'],
-  parameters: {
-    docs: {
-      description: {
-        component: readme,
-      },
-    },
-  },
-} satisfies Meta<typeof Button>;
-
-export default meta;
-
-type Story = StoryObj<typeof meta>;
-
-export const Default: Story = {
-  name: 'Example button',
-};
