+++ packages/storybook/src/documentation/introduction.mdx
@@ -1,6 +1,6 @@
 import { Markdown, Meta } from "@storybook/addon-docs";
 import markdown from "./introduction.md?raw";
 
-<Meta title="Example/Intro" />
+<Meta title="Tilburg/Intro" />
 
 <Markdown>{markdown}</Markdown>
