+++ packages/storybook/src/documentation/license.mdx
@@ -1,6 +1,6 @@
 import { Markdown, Meta } from "@storybook/addon-docs";
 import markdown from "../../../../LICENSE.md?raw";
 
-<Meta title="Example/Open Source License" />
+<Meta title="Tilburg/Open Source License" />
 
 <Markdown>{markdown}</Markdown>
