+++ packages/storybook/src/documentation/notice.mdx
@@ -1,6 +1,6 @@
 import { Markdown, Meta } from "@storybook/addon-docs";
 import markdown from "../../../../NOTICE.md?raw";
 
-<Meta title="Example/Toestemming voor gebruik" />
+<Meta title="Tilburg/Toestemming voor gebruik" />
 
 <Markdown>{markdown}</Markdown>
