+++ proprietary/design-tokens/src/brand/tilburg/border-radius.tokens.json
@@ -1,0 +1,22 @@
+{
+  "tilburg": {
+    "border-radius": {
+      "sm": {
+        "value": "4px",
+        "type": "borderRadius"
+      },
+      "md": {
+        "value": "8px",
+        "type": "borderRadius"
+      },
+      "lg": {
+        "value": "16px",
+        "type": "borderRadius"
+      },
+      "round": {
+        "value": "999px",
+        "type": "borderRadius"
+      }
+    }
+  }
+}
