+++ proprietary/design-tokens/src/brand/tilburg/color.tokens.json
@@ -1,12 +1,195 @@
 {
   "tilburg": {
     "color": {
-      "blue": {
-        "50": { "value": "purple" }
+      "white": {
+        "value": "#ffffff",
+        "type": "color"
       },
-      "grey": {
-        "100": { "value": "white" },
-        "0": { "value": "black" }
+      "black": {
+        "value": "#000000",
+        "type": "color"
+      },
+      "black-txt": {
+        "value": "#333333",
+        "type": "color"
+      },
+      "gray": {
+        "50": {
+          "value": "#F9F9FA",
+          "type": "color"
+        },
+        "100": {
+          "value": "#F2F4F6",
+          "type": "color"
+        },
+        "200": {
+          "value": "#E4E7EC",
+          "type": "color"
+        },
+        "300": {
+          "value": "#CFD5DD",
+          "type": "color"
+        },
+        "400": {
+          "value": "#98A4B6",
+          "type": "color"
+        },
+        "500": {
+          "value": "#7A8AA0",
+          "type": "color"
+        },
+        "600": {
+          "value": "#5B6E8A",
+          "type": "color"
+        },
+        "700": {
+          "value": "#3F5676",
+          "type": "color"
+        },
+        "800": {
+          "value": "#274065",
+          "type": "color"
+        },
+        "900": {
+          "value": "#0A2750",
+          "type": "color"
+        },
+        "950": {
+          "value": "#001737",
+          "type": "color"
+        }
+      },
+      "light": {
+        "alpha": {
+          "50": {
+            "value": "#ffffff0d",
+            "type": "color"
+          },
+          "100": {
+            "value": "#ffffff1a",
+            "type": "color"
+          }
+        }
+      },
+      "dark": {
+        "alpha": {
+          "50": {
+            "value": "#0000000d",
+            "type": "color"
+          },
+          "100": {
+            "value": "#0000001a",
+            "type": "color"
+          }
+        }
+      },
+      "extra": {
+        "yellow": {
+          "value": "#FBDD27",
+          "type": "color"
+        },
+        "purple": {
+          "value": "#513090",
+          "type": "color"
+        },
+        "limegreen": {
+          "value": "#8BC63F",
+          "type": "color"
+        }
+      },
+      "blue": {
+        "100": {
+          "value": "#E6F6FF",
+          "type": "color"
+        },
+        "200": {
+          "value": "#B9E4FC",
+          "type": "color"
+        },
+        "300": {
+          "value": "#00A2E0",
+          "type": "color"
+        },
+        "400": {
+          "value": "#026596",
+          "type": "color"
+        },
+        "500": {
+          "value": "#003865",
+          "type": "color"
+        }
+      },
+      "pink": {
+        "100": {
+          "value": "#FFE8F5",
+          "type": "color"
+        },
+        "200": {
+          "value": "#F5B4D7",
+          "type": "color"
+        },
+        "300": {
+          "value": "#C8006C",
+          "type": "color"
+        },
+        "400": {
+          "value": "#A40058",
+          "type": "color"
+        }
+      },
+      "orange": {
+        "100": {
+          "value": "#FFE7D4",
+          "type": "color"
+        },
+        "200": {
+          "value": "#FBB278",
+          "type": "color"
+        },
+        "300": {
+          "value": "#F16E06",
+          "type": "color"
+        },
+        "400": {
+          "value": "#D45F01",
+          "type": "color"
+        }
+      },
+      "green": {
+        "100": {
+          "value": "#C8F3E9",
+          "type": "color"
+        },
+        "200": {
+          "value": "#86CABB",
+          "type": "color"
+        },
+        "300": {
+          "value": "#008B6D",
+          "type": "color"
+        },
+        "400": {
+          "value": "#01745A",
+          "type": "color"
+        }
+      },
+      "red": {
+        "100": {
+          "value": "#FFE1E0",
+          "type": "color"
+        },
+        "200": {
+          "value": "#F7B0AC",
+          "type": "color"
+        },
+        "300": {
+          "value": "#E5352C",
+          "type": "color"
+        },
+        "400": {
+          "value": "#BF1A12",
+          "type": "color"
+        }
       }
     }
   }
