+++ proprietary/design-tokens/src/brand/tilburg/space.tokens.json
@@ -1,0 +1,253 @@
+{
+  "tilburg": {
+    "space": {
+      "inline": {
+        "flea": {
+          "value": "2px",
+          "type": "spacing"
+        },
+        "ant": {
+          "value": "4px",
+          "type": "spacing"
+        },
+        "beetle": {
+          "value": "8px",
+          "type": "spacing"
+        },
+        "snail": {
+          "value": "12px",
+          "type": "spacing"
+        },
+        "mouse": {
+          "value": "16px",
+          "type": "spacing"
+        },
+        "rat": {
+          "value": "20px",
+          "type": "spacing"
+        },
+        "rabbit": {
+          "value": "24px",
+          "type": "spacing"
+        },
+        "cat": {
+          "value": "28px",
+          "type": "spacing"
+        },
+        "dog": {
+          "value": "32px",
+          "type": "spacing"
+        },
+        "pig": {
+          "value": "48px",
+          "type": "spacing"
+        }
+      },
+      "block": {
+        "flea": {
+          "value": "2px",
+          "type": "spacing"
+        },
+        "ant": {
+          "value": "4px",
+          "type": "spacing"
+        },
+        "beetle": {
+          "value": "8px",
+          "type": "spacing"
+        },
+        "snail": {
+          "value": "12px",
+          "type": "spacing"
+        },
+        "mouse": {
+          "value": "16px",
+          "type": "spacing"
+        },
+        "rat": {
+          "value": "20px",
+          "type": "spacing"
+        },
+        "rabbit": {
+          "value": "24px",
+          "type": "spacing"
+        },
+        "cat": {
+          "value": "32px",
+          "type": "spacing"
+        },
+        "dog": {
+          "value": "48px",
+          "type": "spacing"
+        },
+        "pig": {
+          "value": "64px",
+          "type": "spacing"
+        }
+      },
+      "text": {
+        "flea": {
+          "value": "1px",
+          "type": "spacing",
+          "description": "0.125ch"
+        },
+        "ant": {
+          "value": "2px",
+          "type": "spacing",
+          "description": "0.25ch"
+        },
+        "beetle": {
+          "value": "4px",
+          "type": "spacing",
+          "description": "0.5ch"
+        },
+        "snail": {
+          "value": "6px",
+          "type": "spacing",
+          "description": "0.75ch"
+        },
+        "mouse": {
+          "value": "8px",
+          "type": "spacing",
+          "description": "1ch"
+        },
+        "rat": {
+          "value": "12px",
+          "type": "spacing",
+          "description": "1.5ch"
+        },
+        "rabbit": {
+          "value": "14px",
+          "type": "spacing",
+          "description": "1.75ch"
+        },
+        "cat": {
+          "value": "16px",
+          "type": "spacing",
+          "description": "2ch"
+        },
+        "dog": {
+          "value": "24px",
+          "type": "spacing",
+          "description": "3ch"
+        }
+      },
+      "column": {
+        "flea": {
+          "value": "1px",
+          "type": "spacing"
+        },
+        "ant": {
+          "value": "2px",
+          "type": "spacing"
+        },
+        "beetle": {
+          "value": "4px",
+          "type": "spacing"
+        },
+        "snail": {
+          "value": "8px",
+          "type": "spacing"
+        },
+        "mouse": {
+          "value": "12px",
+          "type": "spacing"
+        },
+        "rat": {
+          "value": "16px",
+          "type": "spacing"
+        },
+        "rabbit": {
+          "value": "20px",
+          "type": "spacing"
+        },
+        "cat": {
+          "value": "24px",
+          "type": "spacing"
+        },
+        "dog": {
+          "value": "28px",
+          "type": "spacing"
+        },
+        "pig": {
+          "value": "32px",
+          "type": "spacing"
+        },
+        "tiger": {
+          "value": "48px",
+          "type": "spacing"
+        },
+        "horse": {
+          "value": "64px",
+          "type": "spacing"
+        },
+        "elephant": {
+          "value": "96px",
+          "type": "spacing"
+        },
+        "giraffe": {
+          "value": "160px",
+          "type": "spacing"
+        }
+      },
+      "row": {
+        "flea": {
+          "value": "1px",
+          "type": "spacing"
+        },
+        "ant": {
+          "value": "2px",
+          "type": "spacing"
+        },
+        "beetle": {
+          "value": "4px",
+          "type": "spacing"
+        },
+        "snail": {
+          "value": "8px",
+          "type": "spacing"
+        },
+        "mouse": {
+          "value": "12px",
+          "type": "spacing"
+        },
+        "rat": {
+          "value": "16px",
+          "type": "spacing"
+        },
+        "rabbit": {
+          "value": "20px",
+          "type": "spacing"
+        },
+        "cat": {
+          "value": "24px",
+          "type": "spacing"
+        },
+        "dog": {
+          "value": "28px",
+          "type": "spacing"
+        },
+        "pig": {
+          "value": "32px",
+          "type": "spacing"
+        },
+        "tiger": {
+          "value": "48px",
+          "type": "spacing"
+        },
+        "horse": {
+          "value": "64px",
+          "type": "spacing"
+        },
+        "elephant": {
+          "value": "96px",
+          "type": "spacing"
+        },
+        "giraffe": {
+          "value": "160px",
+          "type": "spacing"
+        }
+      }
+    }
+  }
+}
