+++ proprietary/design-tokens/src/common/tilburg/interaction.tokens.json
@@ -1,0 +1,22 @@
+{
+  "tilburg": {
+    "interaction": {
+      "color": {
+        "value": "{tilburg.color.pink.300}",
+        "type": "color"
+      },
+      "active": {
+        "color": {
+          "value": "{tilburg.color.pink.200}",
+          "type": "color"
+        }
+      },
+      "hover": {
+        "color": {
+          "value": "{tilburg.color.pink.400}",
+          "type": "color"
+        }
+      }
+    }
+  }
+}
