--- proprietary/design-tokens/src/components/tilburg/button.tokens.json
@@ -1,8 +1,0 @@
-{
-  "tilburg": {
-    "button": {
-      "background-color": { "value": "{tilburg.color.blue.50}" },
-      "color": { "value": "{tilburg.color.grey.100}" }
-    }
-  }
-}
